componentized the shared task list page
